+++ src/Microsoft.AspNetCore.Razor.Evolution/CodeGeneration/IRuntimeTargetExtension.cs
@@ -1,0 +1,9 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+namespace Microsoft.AspNetCore.Razor.Evolution.CodeGeneration
+{
+    public interface IRuntimeTargetExtension
+    {
+    }
+}
