+++ src/Microsoft.AspNetCore.Razor.Evolution/CodeGeneration/TagHelperHtmlAttributeRenderingConventions.cs
@@ -1,0 +1,16 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using Microsoft.AspNetCore.Razor.Evolution.Legacy;
+
+namespace Microsoft.AspNetCore.Razor.Evolution.CodeGeneration
+{
+    internal class TagHelperHtmlAttributeRenderingConventions : CSharpRenderingConventions
+    {
+        public TagHelperHtmlAttributeRenderingConventions(CSharpCodeWriter writer) : base(writer)
+        {
+        }
+
+        public override string StartWriteAttributeValueMethod => "AddHtmlAttributeValue(" /* ORIGINAL: AddHtmlAttributeValueMethodName */;
+    }
+}
