+++ test/Microsoft.AspNetCore.Razor.Evolution.Test/IntegrationTests/CodeGenerationIntegrationTest.cs
@@ -1486,10 +1486,10 @@ namespace Microsoft.AspNetCore.Razor.Evolution.IntegrationTests
                 TestRazorSourceDocument.CreateResource(sourceFilename, encoding: null, normalizeNewLines: true));
 
             // This will ensure that we're not putting any randomly generated data in a baseline.
-            codeDocument.Items[DefaultRazorRuntimeCSharpLoweringPhase.SuppressUniqueIds] = "test";
+            codeDocument.Items[DefaultRazorCSharpLoweringPhase.SuppressUniqueIds] = "test";
 
             // This is to make tests work cross platform.
-            codeDocument.Items[DefaultRazorDesignTimeCSharpLoweringPhase.NewLineString] = "\r\n";
+            codeDocument.Items[DefaultRazorCSharpLoweringPhase.NewLineString] = "\r\n";
 
             return codeDocument;
         }
